Update `more.css' on the front page.
This makes it have everything that was in "scribble.css", so that file
is not brought in here.  (Note the comment about things that probably
should be trimmed out from more.css.)  Even if it's resurrected
later (if it turns out to be more convenient or whatever), it shouldn't
be in common/resources since it's use only on www, and nowhere else.